Add multiple parents check for first request
When a first request is triggered, it is possible to send it for
multiple children. These children must have the same parents.

The application raises an error, since this is not a case that should
happen. There must be an explicit failure in the applications logic to
make this happen.