Add candidate aggregate for registrations
Before each registration belonged to a single child. Now each
registration can be for one or more children. For this purpose the
candidate aggregate is introduced. A candidate is a child, that is
part of a registration. This indirection allows us to keep the child
entity independent of the registration process.